improving the Single Responsibility Principle in the backend code
